requirements.txt
# Core
ib_insync==0.9.86
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Web server (webhooks + dashboard)
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Technical analysis
ta>=0.11.0
ta-lib>=0.4.28; sys_platform != 'win32'

# Market data
yfinance>=0.2.30
alpaca-trade-api>=3.0.0

# Scheduling
apscheduler>=3.10.0

# Notifications
requests>=2.31.0
websocket-client>=1.6.0

# Monitoring
psutil>=5.9.0

# Backtesting
matplotlib>=3.7.0
